requirements.txt
flask
torch
clip-by-openai
Pillow
pinecone-client
